+++ includes/quaternion.h
@@ -6,7 +6,7 @@
 # include "defines.h"
 # include "graceful.h"
 
-t_quaterion	quaternion_new(float angle_x, float angle_y, float angle_z);
+t_quaterion	quaternion_new(double angle_x, double angle_y, double angle_z);
 t_matrix_4	quaternion_to_matrix(t_const_quaterion self);
 
 #endif
