+++ srcs/intersection_common.c
@@ -3,7 +3,7 @@
 void	check_intersection(
 	t_intersection *ret,
 	t_intersect_params *params,
-	float t,
+	double t,
 	t_byte i
 )
 {
